requirements.txt
flask==2.0.1
flask-cors==3.0.10
tensorflow==2.10.0
numpy==1.23.5
pillow==9.0.1
requests==2.28.1
python-dotenv==0.21.0 